refactor cleanup steps into remove hook
